refactor: move interface to models
