+++ src/store/store.ts
@@ -1,4 +1,4 @@
-import { IPlayer, IRoom, IWebSocket } from '../models/models';
+import { gameParams, IPlayer, IRoom, IWebSocket } from '../models/models';
 
 export const Rooms: IRoom[] = [];
 export const Players: IPlayer[] = [];
@@ -8,12 +8,3 @@ export const wssockets: IWebSocket[] = [];
 export const roomId: number = 0;
 export const gameCounter: number = 0;
 export const currentPlayerId: number = 0;
-
-export interface gameParams {
-    roomId: number,
-    wssockets: IWebSocket[],
-    gameCounter: number,
-    firstPlayerShips: any,
-    secondPlayerShips: any,
-    currentPlayerId: number,
-}
